+++ src/common/layers/LayersModule.js
@@ -2,6 +2,7 @@
 
   angular.module('loom_layers', [
     'loom_layers_directive',
-    'loom_layer_info_directive'
+    'loom_layer_info_directive',
+    'loom_layer_attribute_visibility_directive'
   ]);
 })();